Allow command-line/environment configuration
The flask app now looks for a configuration file in the
RQ_DASHBOARD_SETTINGS environment variable and loads it if it is there.

In addition, the rq_dashboard script allows the host and IP of both the
app and the redis server to be specified on the commandline, overriding
the configuration file if necessary, and providing sensible defaults.