Sort list of workers.
The list of worker instances is displayed in non-deterministic order and
changes every time the list refreshes. This makes is hard to follow the
progress of a specific worker instance between AJAX updates.

This change has the workers.json view return the list of worker instances
in a fixed order, based on its state, followed by its name. This way the
busy instances are listed up top.