Do not continue to autoreload the job page if a reload fails.
Autoreload causes the job page to hit the backend every couple of seconds. However, most RQ implementations set a TTL on jobs in redis. So, this polling will eventually fail. If it continues, it will do so continuously. This makes it difficult for users to set up monitoring and alerting, since these errors eventually overwhelm everything else coming from RQ dashboard. And, the refreshes are meaningless anyway.

This PR stops polling on the job page if a reload fails.